should_be_return_12586269025L_when_calculate_given_position_is_50

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/Fibonacci.java
@@ -1,7 +1,7 @@
 package cc.xpbootcamp.warmup.fibonacci;
 
 public class Fibonacci {
-    public static int calculate(int i) {
+    public static long calculate(int i) {
         if (i > 2) {
             return calculate(i - 1) + calculate(i - 2);
         }
